migrate duckdb_load_from_file to duckdb_helpers.py
